Adjust spacing to silence error
In my environment (MacOS Big Sur 11.5.1, zsh shell) executing `Taskfile.sh` throws a number of errors as it searches for `.Taskfile`s:

```
Using: /Users/.Taskfile
/usr/local/bin/t: line 34: /Users/.Taskfile: No such file or directory
/usr/local/bin/t: line 33: [: missing `]'
```

(That's only one; it continues in this vein until it finds a viable `.Taskfile`.)

The script still functions in that it will still display available tasks in the current `.Taskfile` but a screen full of pointless errors is suboptimal. The addition of a single space as per this PR seems to fix the problem.